Move create_edit_hook() to a method of MessageLinker.
This makes usage of it simpler, since you no longer have to call
a function to then immediately call its returned function.
