+++ app/components/github_integration/mentions/fmt.py
@@ -6,12 +6,8 @@ from typing import TYPE_CHECKING
 
 from .cache import entity_cache
 from .resolution import resolve_entity_signatures
-from app.common.hooks import ProcessedMessage
-from app.components.github_integration.models import (
-    Discussion,
-    Issue,
-    PullRequest,
-)
+from app.common.linker import ProcessedMessage
+from app.components.github_integration.models import Discussion, Issue, PullRequest
 from app.utils import dynamic_timestamp, escape_special, format_diff_note
 
 if TYPE_CHECKING:
